Extract shortcut container presentation
